deleted unwanted files and rearrange test files
